+++ packages/components/src/Tooltip/TooltipTrigger.stories.tsx
@@ -1,0 +1,112 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import React, { useState } from 'react';
+
+import { Button } from '../Button';
+import { Tooltip } from './Tooltip';
+
+const meta = {
+  title: 'Components/Tooltip',
+  argTypes: {
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+      table: {
+        type: { summary: 'boolean' },
+        defaultValue: { summary: false },
+      },
+      description: 'If the tooltip is disabled. It will not open.',
+    },
+    delay: {
+      control: {
+        type: 'number',
+      },
+      table: {
+        defaultValue: { summary: '1000' },
+      },
+      description: 'The delay time for the tooltip to show up',
+    },
+    closeDelay: {
+      control: {
+        type: 'number',
+      },
+      table: {
+        defaultValue: { summary: '500' },
+      },
+      description: 'The delay time for the tooltip to close.',
+    },
+    trigger: {
+      control: {
+        type: 'select',
+      },
+      options: ['focus', 'undefined'],
+      table: {
+        defaultValue: { summary: 'focus' },
+      },
+      description:
+        'By default, opens for both focus and hover. Can be made to open only for focus.',
+    },
+    open: {
+      control: {
+        type: 'boolean',
+      },
+      table: {
+        type: { summary: 'boolean' },
+        defaultValue: { summary: false },
+      },
+      description: 'If the tooltip is open',
+    },
+    defaultOpen: {
+      control: {
+        type: 'boolean',
+      },
+      table: {
+        type: { summary: 'boolean' },
+        defaultValue: { summary: false },
+      },
+      description: 'If the tooltip should be default open (uncontrolled)',
+    },
+  },
+} satisfies Meta<typeof Tooltip.Trigger>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Trigger: Story = {
+  render: args => {
+    return (
+      <div className="me-auto ms-auto flex w-[min(100%_-_3rem,60ch)] gap-2 pt-32">
+        <Tooltip.Trigger {...args}>
+          <Button variant="primary">Hover me!</Button>
+          <Tooltip>Look at this tooltip!</Tooltip>
+        </Tooltip.Trigger>
+        <Tooltip.Trigger {...args}>
+          <Button variant="primary">Hover no! Me!</Button>
+          <Tooltip>
+            <div>I am a much more longer tooltip you know!</div>
+            <div>I even have two lines!</div>
+          </Tooltip>
+        </Tooltip.Trigger>
+      </div>
+    );
+  },
+};
+
+export const ControlledTooltipTrigger: Story = {
+  render: args => {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    const [open, setOpen] = useState(false);
+    return (
+      <div className="me-auto ms-auto flex w-[min(100%_-_3rem,60ch)] flex-col gap-2 pt-32">
+        <Tooltip.Trigger open={open} onOpenChange={setOpen} {...args}>
+          <Button variant="primary">Hover Me!</Button>
+          <Tooltip>
+            <div>I am a much more longer tooltip you know!</div>
+            <div>I even have two lines!</div>
+          </Tooltip>
+        </Tooltip.Trigger>
+        <span>Tooltip is {open ? 'showing' : 'not showing'}</span>
+      </div>
+    );
+  },
+};